fix : 주문이 완료되지 않은 경우에만 취소 가능

+++ 2Week_Mission/App/src/main/java/com/ll/exam/final__2022_10_08/app/order/service/OrderService.java
@@ -88,7 +88,9 @@ public class OrderService {
 
   public void cancelOrder(Long id) {
     Order order = orderRepository.findById(id).orElseThrow(null);
-    orderRepository.delete(order);
+    if (order.getReadyStatus().equals(OrderStatus.READY)) {
+      orderRepository.delete(order);
+    }
   }
 
   public Order findOrder(long id) {
